64bit transition part 1: test, .sh.match, init, macro, stk (#867)
This is the first of a thirteen (!!) part patch series that enables
ksh93 to operate within a 64-bit address space (currently dubbed
thickfold). These changes were accomplished by fixing most (but not
all) of the warnings that materialize during compilation when the
following clang compiler flags are passed: '-Wsign-compare
-Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion'

Many ints are replaced with ssize_t. While POSIX does not guarantee
that ssize_t will accept any negative value besides -1, it also
does not guarantee it will accept values > 32767, so this commit is
going with real-world portability. The assumption that ssize_t is
the signed version of size_t and will accept values in the range
-(SSIZE_MAX+1)...SSIZE_MAX works fine on every known system that
ksh runs on. Some ast/ksh code was already depending on this.

In some areas ssize_t results from e.g. sfvalue() may end up being
used with ptrdiff_t variables. This is not pedantically correct,
but it's certainly better than the prior int hell status quo,
wherein ssize_t was usually shortened to int.

Conspicuous changes with noteworthyness:
- Fixed many, many compiler warnings by adding a considerable
  number of casts and changing the types of more than quite
  a few variables. This is a consequence of using compiler
  warnings as an aid for implementing 64-bit memory allocation;
  while the warnings were useful, there were so many warnings
  fixed in the process that it increased the patch sizes greatly.
- Transitioned the strgrpmatch() and strngrpmatch calls to use
  a ssize_t* pointer rather than an int* pointer.
- For the sh_options macros/function, enforce uint64_t as the main
  argument type to fix compiler warnings.
- The libast stk sublibrary already uses the ssize_t/size_t types,
  which made it rather easy to integrate proper ptrdiff_t usage.
  In fact, stktell() has *always* returned a 'ptrdiff_t' result.
  The documentation in stk(3) has been incorrectly claiming
  since < 1995 it returns 'int', which is wrong. That error
  has been rectified alongside the other updates to stk.

The thickfold patch series was accomplished by fixing compiler
warnings, so I've decided to provide metrics as to the change in
the total number of warnings occurring during compilation.
Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 4,042 => 3,808 => 248 (progression from
cc5e0692 => part 1 => part 13)

Side note: To re-emphasize, this patch is one part of a whole
(although it can be used on it's own). The full suite of changes
can be found on the thickfold-size_t branch. This first part has
been submitted severed from the other changes to make code review
less laborious (I hope). The full patch series can be found here:
https://github.com/JohnoKing/ksh/tree/ad588c003d/patches

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/cmd/ksh93/include/defs.h
@@ -115,7 +115,7 @@ extern pid_t		_sh_fork(pid_t, int ,int*);
 extern void		sh_invalidate_ifs(void);
 extern char 		*sh_mactrim(char*,int);
 extern int 		sh_macexpand(struct argnod*,struct argnod**,int);
-extern int		sh_macfun(const char*,int);
+extern int		sh_macfun(const char*,ptrdiff_t);
 extern void 		sh_machere(Sfio_t*, Sfio_t*, char*);
 extern void 		*sh_macopen(void);
 extern char 		*sh_macpat(struct argnod*,int);
@@ -126,7 +126,7 @@ extern int		sh_mathstd(const char*);
 extern void		sh_printopts(Shopt_t,int,Shopt_t*);
 extern int 		sh_readline(char**,volatile int,int,ssize_t,Sflong_t);
 extern Sfio_t		*sh_sfeval(char*[]);
-extern void		sh_setmatch(const char*,int,int,int[],int);
+extern void		sh_setmatch(const char*,ptrdiff_t,ssize_t,ssize_t[],int);
 extern void		sh_scope(struct argnod*, int);
 extern Namval_t		*sh_scoped(Namval_t*);
 extern Dt_t		*sh_subtracktree(int);
@@ -158,7 +158,7 @@ extern void		*sh_calloc(size_t nmemb, size_t size);
 extern char		*sh_strdup(const char *s);
 extern void		*sh_memdup(const void *s, size_t n);
 extern char		*sh_getcwd(void);
-#define new_of(type,x)	((type*)sh_malloc((unsigned)sizeof(type)+(x)))
+#define new_of(type,x)	((type*)sh_malloc(sizeof(type)+(x)))
 #define sh_newof(p,t,n,x)	((p)?(t*)sh_realloc((char*)(p),sizeof(t)*(n)+(x)):(t*)sh_calloc(1,sizeof(t)*(n)+(x)))
 
 #define URI_RFC3986_UNRESERVED "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~"
@@ -173,7 +173,7 @@ extern char		*sh_getcwd(void);
 
 #if SHOPT_SCRIPTONLY
 #define is_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] & ((uint64_t)1 << ((x) % WBITS))) )
-#define on_option(s,x)	( (x)==SH_INTERACTIVE || (x)==SH_HISTORY ? errormsg(SH_DICT,ERROR_exit(1),e_scriptonly) : ((s)->v[((x)&WMASK)/WBITS] |= ((uint64_t)1 << ((x) % WBITS))) )
+#define on_option(s,x)	( (x)==SH_INTERACTIVE || (x)==SH_HISTORY ? (uint64_t)errormsg(SH_DICT,ERROR_exit(1),e_scriptonly) : ((s)->v[((x)&WMASK)/WBITS] |= ((uint64_t)1 << ((x) % WBITS))) )
 #define off_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] &= ~((uint64_t)1 << ((x) % WBITS))) )
 #else
 #define is_option(s,x)	((s)->v[((x)&WMASK)/WBITS] & ((uint64_t)1 << ((x) % WBITS)))
